raspi: fix 32/64-bit pointer/integer warnings in shared code
Route pointer<->integer conversions through IPTR (uintptr_t in boot
code) in usb2otg, vc4gfx and the aarch64 boot loader, so the casts
are correct on both 32-bit arm and 64-bit aarch64 builds. Use ~0U for
32-bit register writes, cast bug() arguments to match %lx formats,
drop the VCPOWER_STATE_* duplicates of hardware/videocore.h, and fold
debug-only variables in elf.c into their DELF() blocks.

+++ arch/aarch64-raspi/boot/vc_fb.c
@@ -119,7 +119,7 @@ int vcfb_init(void)
         D(kprintf("%p, %p, %p, %p, %p\n", AROS_LE2LONG(vcmb_msg[count]), AROS_LE2LONG(vcmb_msg[count+1]), AROS_LE2LONG(vcmb_msg[count+2]),
         AROS_LE2LONG(vcmb_msg[count+3]),AROS_LE2LONG(vcmb_msg[count+4])));
 
-        if (((scr_FrameBuffer = (void *)(AROS_LE2LONG(vcmb_msg[count + 3]))) == 0) || (AROS_LE2LONG(vcmb_msg[count + 4]) == 0))
+        if (((scr_FrameBuffer = (void *)(uintptr_t)AROS_LE2LONG(vcmb_msg[count + 3])) == 0) || (AROS_LE2LONG(vcmb_msg[count + 4]) == 0))
             return 0;
 
         D(kprintf("[VCFB] scr_Framebuffer=%p, %p\n", scr_FrameBuffer, (intptr_t)scr_FrameBuffer & 0xc0000000));
